requirements.txt
numpy==1.24.3
opencv_python==4.8.0.74
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.3.0
skimage==0.21.0
tensorflow==2.12.1
